refactor: split renderer config and extract particle trails helpers
