+++ .prettierignore
@@ -7,3 +7,4 @@ playwright-report
 tmp
 *.min.js
 *.bundle.js
+temp-lint-main.json
