+++ test/vitest/shield-bubble-visibility.spec.tsx
@@ -7,7 +7,7 @@ import { SHIELD_TUNING, SHIELD_VISUALS } from '../../src/config/renderer.js';
 
 describe('ShieldBubble visibility behavior (static analysis)', () => {
   const shipFilePath = path.resolve(__dirname, '../../src/components/ship/ShipShield.tsx');
-  const configFilePath = path.resolve(__dirname, '../../src/config/renderer.ts');
+  const configFilePath = path.resolve(__dirname, '../../src/config/effects.ts');
 
   it('includes conditional rendering for low shield levels', () => {
     const txt = fs.readFileSync(shipFilePath, 'utf-8');
@@ -27,7 +27,7 @@ describe('ShieldBubble visibility behavior (static analysis)', () => {
 
   it('fixes HULL_TINT threshold to prevent always-on hull tinting', () => {
     const txt = fs.readFileSync(configFilePath, 'utf-8');
-    
+
     // HULL_TINT should have a low threshold, not 1.00 which would always apply
     expect(txt).toContain('tintThreshold: 0.02');
     expect(txt).not.toContain('tintThreshold: 1.00');
@@ -43,22 +43,22 @@ describe('ShieldBubble visibility behavior (static analysis)', () => {
 
     // With 0 shields and any maxShield > 0, the fraction should be 0
     // And 0 < minShieldThreshold (0.01) should be true, causing early return
-    const shieldFraction = 0 / Math.max(1, 100);  // 0
+    const shieldFraction = 0 / Math.max(1, 100); // 0
     const minThreshold = 0.01;
     expect(shieldFraction < minThreshold).toBe(true);
   });
 
   it('ensures shield bubble logic allows rendering above threshold', () => {
     const txt = fs.readFileSync(shipFilePath, 'utf-8');
-    
+
     // Test that shields above 1% should render
-    const shieldFractionFull = 100 / Math.max(1, 100);  // 1.0
-    const shieldFractionPartial = 50 / Math.max(1, 100);  // 0.5  
-    const shieldFractionMinimal = 1 / Math.max(1, 100);   // 0.01
+    const shieldFractionFull = 100 / Math.max(1, 100); // 1.0
+    const shieldFractionPartial = 50 / Math.max(1, 100); // 0.5
+    const shieldFractionMinimal = 1 / Math.max(1, 100); // 0.01
     const minThreshold = 0.01;
-    
-    expect(shieldFractionFull >= minThreshold).toBe(true);   // 100% shields should render
-    expect(shieldFractionPartial >= minThreshold).toBe(true); // 50% shields should render  
+
+    expect(shieldFractionFull >= minThreshold).toBe(true); // 100% shields should render
+    expect(shieldFractionPartial >= minThreshold).toBe(true); // 50% shields should render
     expect(shieldFractionMinimal >= minThreshold).toBe(true); // 1% shields should render
   });
 
@@ -91,7 +91,8 @@ describe('ShieldBubble visibility behavior (static analysis)', () => {
 
       // When opacity is low/zero, the shader guarantees a minimum visible floor:
       // uOpacity * uMaxAlpha * uMinAlphaFloor should be the minimum alpha contributed
-      const floorAlpha = uniforms.uOpacity.value * uniforms.uMaxAlpha.value * uniforms.uMinAlphaFloor.value;
+      const floorAlpha =
+        uniforms.uOpacity.value * uniforms.uMaxAlpha.value * uniforms.uMinAlphaFloor.value;
       expect(floorAlpha).toBeGreaterThan(0);
 
       mat.dispose();
